Updated Waiter to use abstract class with concrete implementation.
The Waiter now uses an abstract class to define basic structure, and its
concrete implementations may define several ways of running the
try/retry/timeout pattern.

+++ src/main/java/org/atom/quark/sftp/helper/SftpHelper.java
@@ -235,7 +235,7 @@ public interface SftpHelper extends Helper {
 	 * @throws Exception 
 	 * @throws SftpException 
 	 */
-	public TypedHelperResult<Vector<LsEntry>> waitForContainsFile(String dir, Pattern pattern, long timeout, int period) throws Exception;
+	public TypedHelperResult<Vector<LsEntry>> waitForContainsFile(String dir, Pattern pattern, long timeout, long period) throws Exception;
 	
 	/**
 	 * Check whether a directory contains files matching the given pattern. 
@@ -262,7 +262,7 @@ public interface SftpHelper extends Helper {
 	 * @throws SftpException 
 	 */
 	public TypedHelperResult<Vector<LsEntry>> waitForContainsFile(String dir, Pattern pattern, int count,
-			long timeout, int period) throws Exception;
+			long timeout, long period) throws Exception;
 	
 	/**
 	 * Retrieve the MD5 checksum for the given file on SFTP server.
